[quant] Move EmbeddingBag eager quantization to static
Summary:
Move the tests to static ones as well

Test Plan:
 python test/test_quantization.py TestStaticQuantizedModule.test_embedding_bag_api

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
