+++ test/quantization/test_quantize.py
@@ -528,6 +528,49 @@ class TestPostTrainingStatic(QuantizationTestCase):
         self.assertTrue('QuantizedLinear' in str(model))
         self.checkQuantizedLinear(model.fc)
 
+
+    def test_quantized_embedding_bag(self):
+        r""" Test the post-training quantization flow, serialization and scripting
+        of embedding_bag modules
+        """
+        model = EmbeddingBagModule().eval()
+        indices = torch.tensor([9, 6, 5, 7, 8, 8, 9, 2, 8, 6, 6, 9, 1, 6, 8, 8, 3, 2, 3, 6, 3, 6, 5, 7, 0, 8, 4, 6, 5, 8, 2, 3])
+        offsets = torch.tensor([0, 19, 20, 28, 28, 32])
+        weights = torch.randn(10, 12, dtype=torch.float32)
+
+        model.qconfig = float_qparams_dynamic_qconfig
+        prepare(model, inplace=True)
+        quantized_model = convert(model)
+
+        per_sample_weights = torch.from_numpy(np.random.uniform(
+            low=0.01, high=0.5, size=[len(indices)]).astype(np.float32))
+
+        # Test to make sure module is quantized correctly.
+        self.assertTrue('QuantizedEmbeddingBag' in str(quantized_model))
+        self.checkDynamicQuantizedModule(quantized_model.emb, torch.nn.quantized.EmbeddingBag, torch.quint8)
+        self.checkScriptable(quantized_model, [[indices, offsets, per_sample_weights]], check_save_load=True)
+
+        class EmbeddingBagWithLinear(torch.nn.Module):
+            def __init__(self):
+                super().__init__()
+                self.emb = torch.nn.EmbeddingBag(num_embeddings=10, embedding_dim=12,
+                                                 include_last_offset=True, scale_grad_by_freq=False, mode='sum')
+                self.fc = torch.nn.Linear(5, 5)
+
+            def forward(self, indices, offsets, per_sample_weights, linear_in):
+                return self.emb(indices, offsets, per_sample_weights), self.fc(linear_in)
+
+        # Test quantization of embedding_bag layer only
+        model = EmbeddingBagWithLinear().eval()
+        model.emb.qconfig = float_qparams_dynamic_qconfig
+        prepare(model, inplace=True)
+        quantized_model = convert(model)
+
+        self.assertTrue('QuantizedEmbeddingBag' in str(quantized_model))
+        self.checkLinear(model.fc)
+        self.checkDynamicQuantizedModule(quantized_model.emb, torch.nn.quantized.EmbeddingBag, torch.quint8)
+
+
 @skipIfNoFBGEMM
 class TestPostTrainingDynamic(QuantizationTestCase):
     def test_single_layer(self):
@@ -883,42 +926,6 @@ class TestPostTrainingDynamic(QuantizationTestCase):
             convert_dynamic(model)
             checkHooksIsPresent(model)
 
-    def test_quantized_embedding_bag(self):
-        r""" Test the post-training dynamic quantization flow, serialization and scripting
-        of embedding_bag modules
-        """
-        model = EmbeddingBagModule().eval()
-        indices = torch.tensor([9, 6, 5, 7, 8, 8, 9, 2, 8, 6, 6, 9, 1, 6, 8, 8, 3, 2, 3, 6, 3, 6, 5, 7, 0, 8, 4, 6, 5, 8, 2, 3])
-        offsets = torch.tensor([0, 19, 20, 28, 28, 32])
-        weights = torch.randn(10, 12, dtype=torch.float32)
-
-        quantized_model = quantize_dynamic(model, dtype=torch.quint8)
-
-        per_sample_weights = torch.from_numpy(np.random.uniform(
-            low=0.01, high=0.5, size=[len(indices)]).astype(np.float32))
-
-        # Test to make sure module is quantized correctly.
-        self.assertTrue('DynamicQuantizedEmbeddingBag' in str(quantized_model))
-        self.checkDynamicQuantizedModule(quantized_model.emb, torch.nn.quantized.dynamic.EmbeddingBag, torch.quint8)
-        self.checkScriptable(quantized_model, [[indices, offsets, per_sample_weights]], check_save_load=True)
-
-        class EmbeddingBagWithLinear(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.emb = torch.nn.EmbeddingBag(num_embeddings=10, embedding_dim=12,
-                                                 include_last_offset=True, scale_grad_by_freq=False, mode='sum')
-                self.fc = torch.nn.Linear(5, 5)
-
-            def forward(self, indices, offsets, per_sample_weights, linear_in):
-                return self.emb(indices, offsets, per_sample_weights), self.fc(linear_in)
-
-        # Test set API
-        model = EmbeddingBagWithLinear().eval()
-        quantized_model = quantize_dynamic(model, set([nn.EmbeddingBag]), dtype=torch.quint8)
-
-        self.assertTrue('DynamicQuantizedEmbeddingBag' in str(quantized_model))
-        self.checkLinear(model.fc)
-        self.checkDynamicQuantizedModule(quantized_model.emb, torch.nn.quantized.dynamic.EmbeddingBag, torch.quint8)
 
 
 class TestQuantizationAwareTraining(QuantizationTestCase):
