+++ torch/nn/quantized/modules/__init__.py
@@ -8,7 +8,7 @@ from .normalization import LayerNorm, GroupNorm, InstanceNorm1d, \
     InstanceNorm2d, InstanceNorm3d
 from .conv import Conv1d, Conv2d, Conv3d
 from .linear import Linear
-from .embedding import Embedding
+from .embedding_ops import Embedding, EmbeddingBag
 
 from .functional_modules import FloatFunctional, QFunctional
 
@@ -99,6 +99,7 @@ __all__ = [
     'InstanceNorm2d',
     'InstanceNorm3d',
     'Embedding',
+    'EmbeddingBag',
     # Wrapper modules
     'FloatFunctional',
     'QFunctional',
