+++ torch/quantization/default_mappings.py
@@ -29,6 +29,7 @@ DEFAULT_MODULE_MAPPING = {
     nn.InstanceNorm2d: nnq.InstanceNorm2d,
     nn.InstanceNorm3d: nnq.InstanceNorm3d,
     nn.Embedding: nnq.Embedding,
+    nn.EmbeddingBag: nnq.EmbeddingBag,
     QuantStub: nnq.Quantize,
     DeQuantStub: nnq.DeQuantize,
     # Wrapper Modules:
@@ -75,7 +76,6 @@ DEFAULT_DYNAMIC_MODULE_MAPPING = {
     nn.LSTMCell: nnqd.LSTMCell,
     nn.RNNCell: nnqd.RNNCell,
     nn.GRUCell: nnqd.GRUCell,
-    nn.EmbeddingBag: nnqd.EmbeddingBag,
 }
 
 # Allowed list for propagating the qconfig
